# house_app_server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for house management including image and file storage.",
        "title": "House App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/images": {
            "get": {
                "description": "Get a list of all images",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Get all images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Image"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Upload a single image file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Upload a single image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tags",
                        "name": "tags",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Image"
                        }
                    }
                }
            }
        },
        "/images/bulk-upload": {
            "post": {
                "description": "Upload multiple images from a specified folder path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Bulk upload images from a source folder",
                "parameters": [
                    {
                        "description": "Bulk upload request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BulkUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BulkUploadResponse"
                        }
                    }
                }
            }
        },
        "/images/multiple": {
            "post": {
                "description": "Upload multiple image files at once from a mobile device or web browser",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Upload multiple images",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image files to upload (select multiple)",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tags for all images",
                        "name": "tags",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BulkUploadResponse"
                        }
                    }
                }
            }
        },
        "/images/search": {
            "get": {
                "description": "Search for images by filename or tags. Requires a text index on the collection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Search images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Image"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BulkUploadRequest": {
            "type": "object",
            "properties": {
                "sourceFolder": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.BulkUploadResponse": {
            "type": "object",
            "properties": {
                "failedCount": {
                    "type": "integer"
                },
                "failedFiles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "successfulCount": {
                    "type": "integer"
                },
                "totalImages": {
                    "type": "integer"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "contentType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {},
                "originalPath": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "storagePath": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}